Moving and renaming some components + create a main container around all routes
