+++ src/App.tsx
@@ -4,8 +4,8 @@ import { ThemeProvider } from '@iexec/react-ui-kit';
 import {
   Navigation,
   NewProtectedData,
-  ProtectedData,
-  Consent,
+  MyProtectedData,
+  OneProtectedData,
   EmailDapp,
   SendEmail,
   LoginGuard,
@@ -23,50 +23,52 @@ function App() {
     <div className="App">
       <ThemeProvider>
         <Navigation />
-        <Routes>
-          <Route
-            path={`/${PROTECTED_DATA}`}
-            element={
-              <LoginGuard>
-                <ProtectedData />
-              </LoginGuard>
-            }
-          />
-          <Route
-            path={`/${PROTECTED_DATA}/${CONSENT}/:ProtectedDataId`}
-            element={
-              <LoginGuard>
-                <Consent />
-              </LoginGuard>
-            }
-          />
-          <Route
-            path={`/${PROTECTED_DATA}/${CREATE}`}
-            element={
-              <LoginGuard>
-                <NewProtectedData />
-              </LoginGuard>
-            }
-          />
-          <Route
-            path={`/${SEND_EMAIL}`}
-            element={
-              <LoginGuard>
-                <EmailDapp />
-              </LoginGuard>
-            }
-          />
-          <Route
-            path={`/${SEND_EMAIL}/:receiverAddress/:protectedDataAddress`}
-            element={
-              <LoginGuard>
-                <SendEmail />
-              </LoginGuard>
-            }
-          />
-          {/* default redirect */}
-          <Route path="*" element={<Navigate to={`/${HOME}`} />} />
-        </Routes>
+        <div className="mx-auto mt-12 w-[70%]">
+          <Routes>
+            <Route
+              path={`/${PROTECTED_DATA}`}
+              element={
+                <LoginGuard>
+                  <MyProtectedData />
+                </LoginGuard>
+              }
+            />
+            <Route
+              path={`/${PROTECTED_DATA}/${CONSENT}/:ProtectedDataId`}
+              element={
+                <LoginGuard>
+                  <OneProtectedData />
+                </LoginGuard>
+              }
+            />
+            <Route
+              path={`/${PROTECTED_DATA}/${CREATE}`}
+              element={
+                <LoginGuard>
+                  <NewProtectedData />
+                </LoginGuard>
+              }
+            />
+            <Route
+              path={`/${SEND_EMAIL}`}
+              element={
+                <LoginGuard>
+                  <EmailDapp />
+                </LoginGuard>
+              }
+            />
+            <Route
+              path={`/${SEND_EMAIL}/:receiverAddress/:protectedDataAddress`}
+              element={
+                <LoginGuard>
+                  <SendEmail />
+                </LoginGuard>
+              }
+            />
+            {/* default redirect */}
+            <Route path="*" element={<Navigate to={`/${HOME}`} />} />
+          </Routes>
+        </div>
       </ThemeProvider>
     </div>
   );
