+++ src/features/index.ts
@@ -1,7 +1,7 @@
 export { default as LoginGuard } from './loginGuard/LoginGuard';
 export { default as Navigation } from './navigation/Navigation';
-export { default as ProtectedData } from './protectedData/ProtectedData';
-export { default as NewProtectedData } from './newProtectedData/NewProtectedData';
-export { default as Consent } from './consent/Consent';
+export { default as MyProtectedData } from './myProtectedData/MyProtectedData';
+export { default as NewProtectedData } from './myProtectedData/newProtectedData/NewProtectedData';
+export { default as OneProtectedData } from './myProtectedData/OneProtectedData/OneProtectedData';
 export { default as EmailDapp } from './emailDapp/EmailDapp';
 export { default as SendEmail } from './sendEmail/SendEmail';
