+++ src/features/myProtectedData/OneProtectedData/OneProtectedData.css
@@ -1,9 +1,3 @@
-#consent {
-  margin: auto;
-  margin-top: 90px;
-  width: 70%;
-}
-
 .chipType {
   margin-top: 10px;
   background-color: #fcd15a;
