+++ src/features/myProtectedData/newProtectedData/NewProtectedData.css
@@ -1,0 +1,4 @@
+.uploadFileButton {
+  margin-top: 20px;
+  width: 100%;
+}
