+++ src/features/navigation/Navigation.tsx
@@ -2,7 +2,7 @@ import { NavBar } from '@iexec/react-ui-kit';
 import { useNavigate, useMatch } from 'react-router-dom';
 import { useAccount, useDisconnect } from 'wagmi';
 import { useWeb3Modal } from '@web3modal/react';
-import { resetAppState, selectAppIsConnected } from '../../app/appSlice';
+import { resetAppState, selectAppIsConnected } from '@/app/appSlice';
 import { useAppDispatch, useAppSelector } from '../../app/hooks';
 import { HOME, PROTECTED_DATA, SEND_EMAIL } from '../../config/path';
 import { bellecour } from '../../utils/walletConnection';
