--- src/features/newProtectedData/NewProtectedData.css
@@ -1,14 +1,0 @@
-#newProtectedData {
-  margin: auto;
-  margin-top: 90px;
-  width: 70%;
-}
-
-.uploadFileButton {
-  margin-top: 20px;
-  width: 100%;
-}
-
-.protectNewDataButton {
-  margin-top: 20px;
-}
